+++ docs/changelog.md
@@ -43,6 +43,8 @@ What's changed since pre-release v1.48.0-B0088:
     [#3649](https://github.com/Azure/PSRule.Rules.Azure/issues/3649)
   - Added validation for Bicep custom type constraints.
     [#1922](https://github.com/Azure/PSRule.Rules.Azure/issues/1922)
+  - Added support for resolving filtered subnet IDs from existing virtual networks during Bicep expansion.
+    [#2159](https://github.com/Azure/PSRule.Rules.Azure/issues/2159)
 
 ## v1.48.0-B0088 (pre-release)